+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Development
 
 - Restapi: Unify station parameter and add alias
 - Interpolation: Make maximum station distance per parameter configurable via settings
+- Result: Convert date to string only if dataframe is not empty
 
 0.68.0 (01.12.2023)
 *******************