+++ CHANGELOG.rst
@@ -4,6 +4,9 @@ Changelog
 Development
 ***********
 
+0.69.0 (18.12.2023)
+*******************
+
 - Restapi: Unify station parameter and add alias
 - Interpolation: Make maximum station distance per parameter configurable via settings
 - Result: Convert date to string only if dataframe is not empty